+++ sdk/messaging/azservicebus/liveTestHelpers_test.go
@@ -40,6 +40,16 @@ func getConnectionString(t *testing.T) string {
 	return cs
 }
 
+func getConnectionStringWithoutManagePerms(t *testing.T) string {
+	cs := os.Getenv("SERVICEBUS_CONNECTION_STRING_NO_MANAGE")
+
+	if cs == "" {
+		t.Skip()
+	}
+
+	return cs
+}
+
 // createQueue creates a queue using a subset of entries in 'queueDescription':
 // - EnablePartitioning
 // - RequiresSession