+++ sdk/resourcemanager/databricks/armdatabricks/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armdatabricks
 
 const (
 	module  = "armdatabricks"
-	version = "v0.2.0"
+	version = "v0.2.1"
 )
 
 // CreatedByType - The type of identity that created the resource.