+++ sdk/storage/azblob/constants.go
@@ -30,4 +30,6 @@ const (
 	rfc3339   = "2006-01-02T15:04:05.9999999Z"
 	timestamp = "Timestamp"
 	etag      = "ETag"
+
+	tokenScope = "https://storage.azure.com/.default"
 )