+++ sdk/storage/azblob/zc_shared_policy_shared_key_credential.go
@@ -17,7 +17,6 @@ import (
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/log"
 )
 
@@ -166,12 +165,8 @@ type sharedKeyCredPolicy struct {
 	cred *SharedKeyCredential
 }
 
-func newSharedKeyCredPolicy(cred *SharedKeyCredential, opts runtime.AuthenticationOptions) *sharedKeyCredPolicy {
-	s := &sharedKeyCredPolicy{
-		cred: cred,
-	}
-
-	return s
+func newSharedKeyCredPolicy(cred *SharedKeyCredential) *sharedKeyCredPolicy {
+	return &sharedKeyCredPolicy{cred: cred}
 }
 
 func (s *sharedKeyCredPolicy) Do(req *policy.Request) (*http.Response, error) {
@@ -192,12 +187,7 @@ func (s *sharedKeyCredPolicy) Do(req *policy.Request) (*http.Response, error) {
 	response, err := req.Next()
 	if err != nil && response != nil && response.StatusCode == http.StatusForbidden {
 		// Service failed to authenticate request, log it
-		log.Write(log.Response, "===== HTTP Forbidden status, String-to-NewSASQueryParameters:\n"+stringToSign+"\n===============================\n")
+		log.Write(log.EventResponse, "===== HTTP Forbidden status, String-to-NewSASQueryParameters:\n"+stringToSign+"\n===============================\n")
 	}
 	return response, err
-}
-
-// NewAuthenticationPolicy implements the Credential interface on SharedKeyCredential.
-func (c *SharedKeyCredential) NewAuthenticationPolicy(options runtime.AuthenticationOptions) policy.Policy {
-	return newSharedKeyCredPolicy(c, options)
 }