+++ sdk/azcore/CHANGELOG.md
@@ -1,12 +1,13 @@
 # Release History
 
-## 0.20.1 (Unreleased)
+## 0.21.0 (Unreleased)
 
 ### Features Added
 * Added `AllowedHeaders` and `AllowedQueryParams` to `policy.LogOptions` to control which headers and query parameters are written to the logger.
 
 ### Breaking Changes
 * Moved `[]policy.Policy` parameters of `arm/runtime.NewPipeline` and `runtime.NewPipeline` into a new struct, `runtime.PipelineOptions`
+* Renamed `arm/ClientOptions.Host` to `.Endpoint`
 
 ### Bugs Fixed